+++ services/flight-service/src/main/java/com/boeing/flightservice/service/impl/RouteServiceImpl.java
@@ -1,13 +1,5 @@
 package com.boeing.flightservice.service.impl;
 
-import java.util.Map;
-import java.util.Objects;
-import java.util.UUID;
-
-import org.springframework.http.converter.json.MappingJacksonValue;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import com.boeing.flightservice.dto.paging.RouteDto;
 import com.boeing.flightservice.dto.request.RouteCreateRequestDTO;
 import com.boeing.flightservice.dto.response.RouteResponseDTO;
@@ -19,14 +11,20 @@ import com.boeing.flightservice.repository.AirportRepository;
 import com.boeing.flightservice.repository.RouteRepository;
 import com.boeing.flightservice.service.spec.RouteService;
 import com.boeing.flightservice.util.PaginationUtil;
-
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.converter.json.MappingJacksonValue;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Map;
+import java.util.Objects;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
 @Transactional
 public class RouteServiceImpl implements RouteService {
-    
+
     private final RouteRepository routeRepository;
     private final AirportRepository airportRepository;
 
@@ -41,7 +39,6 @@ public class RouteServiceImpl implements RouteService {
 
     @Override
     public RouteResponseDTO createRoute(RouteCreateRequestDTO request) {
-        // Validate that origin and destination are different
         if (Objects.equals(request.getOriginAirportId(), request.getDestinationAirportId())) {
             throw new BadRequestException("Origin and destination airports cannot be the same");
         }
@@ -49,7 +46,7 @@ public class RouteServiceImpl implements RouteService {
         // Find airports
         Airport origin = airportRepository.findByIdAndDeleted(request.getOriginAirportId(), false)
                 .orElseThrow(() -> new ResourceNotFoundException("Origin airport not found"));
-        
+
         Airport destination = airportRepository.findByIdAndDeleted(request.getDestinationAirportId(), false)
                 .orElseThrow(() -> new ResourceNotFoundException("Destination airport not found"));
 
@@ -63,20 +60,22 @@ public class RouteServiceImpl implements RouteService {
                 .origin(origin)
                 .destination(destination)
                 .estimatedDurationMinutes(request.getEstimatedDurationMinutes())
-                .build();        route = routeRepository.save(route);
-        
+                .build();
+        route = routeRepository.save(route);
+
         return RouteResponseDTO.fromEntity(route);
     }
 
     @Override
-    public RouteResponseDTO getRouteById(UUID id) {        Route route = routeRepository.findByIdAndDeleted(id, false)
+    public RouteResponseDTO getRouteById(UUID id) {
+        Route route = routeRepository.findByIdAndDeleted(id, false)
                 .orElseThrow(() -> new ResourceNotFoundException("Route not found"));
-        
+
         return RouteResponseDTO.fromEntity(route);
     }
 
     @Override
-    public Double getEstimatedDuration(UUID originAirportId, UUID destinationAirportId) {
+    public Integer getEstimatedDuration(UUID originAirportId, UUID destinationAirportId) {
         return routeRepository.findByOriginAndDestination(originAirportId, destinationAirportId)
                 .map(Route::getEstimatedDurationMinutes)
                 .orElse(null);
